requirements.txt
discord.py==1.6.0
PyYAML==5.3.1
